Overhaul the event model, adopt document-based model
UI code isn't going to be re-used, so it's allowed to interface
with the model directly. The model code is being re-used, so
it shouldn't know about the view.

So I flipped the events around. Instead of the view producing
events and the model consuming them, the model produces events
and the view consumes them.

I also adopted a document-based approach.